// gauge-ts/src/gen/messages.ts
export interface ParameterPosition {
  oldPosition: number;
  newPosition: number;
}

/** / Tells the runner to refactor the specified Step. */
export interface RefactorRequest {
  /** / Old value, used to lookup Step to refactor */
  oldStepValue:
    | ProtoStepValue
    | undefined;
  /** / New value, the to-be value of Step being refactored. */
  newStepValue:
    | ProtoStepValue
    | undefined;
  /** / Holds parameter positions of all parameters. Contains old and new parameter positions. */
  paramPositions: ParameterPosition[];
  /** / If set to true, the refactored files should be saved to the file system before returning the response. */
  saveChanges: boolean;
}
// ...
export const ParameterPosition: MessageFns<ParameterPosition> = {
  encode(message: ParameterPosition, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.oldPosition !== 0) {
      writer.uint32(8).int32(message.oldPosition);
    }
    if (message.newPosition !== 0) {
      writer.uint32(16).int32(message.newPosition);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ParameterPosition {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseParameterPosition();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.oldPosition = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.newPosition = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ParameterPosition {
    return {
      oldPosition: isSet(object.oldPosition) ? globalThis.Number(object.oldPosition) : 0,
      newPosition: isSet(object.newPosition) ? globalThis.Number(object.newPosition) : 0,
    };
  },

  toJSON(message: ParameterPosition): unknown {
    const obj: any = {};
    if (message.oldPosition !== 0) {
      obj.oldPosition = Math.round(message.oldPosition);
    }
    if (message.newPosition !== 0) {
      obj.newPosition = Math.round(message.newPosition);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ParameterPosition>, I>>(base?: I): ParameterPosition {
    return ParameterPosition.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ParameterPosition>, I>>(object: I): ParameterPosition {
    const message = createBaseParameterPosition();
    message.oldPosition = object.oldPosition ?? 0;
    message.newPosition = object.newPosition ?? 0;
    return message;
  },
};

function createBaseRefactorRequest(): RefactorRequest {
  return { oldStepValue: undefined, newStepValue: undefined, paramPositions: [], saveChanges: false };
}

export const RefactorRequest: MessageFns<RefactorRequest> = {
  encode(message: RefactorRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.oldStepValue !== undefined) {
      ProtoStepValue.encode(message.oldStepValue, writer.uint32(10).fork()).join();
    }
    if (message.newStepValue !== undefined) {
      ProtoStepValue.encode(message.newStepValue, writer.uint32(18).fork()).join();
    }
    for (const v of message.paramPositions) {
      ParameterPosition.encode(v!, writer.uint32(26).fork()).join();
    }
    if (message.saveChanges !== false) {
      writer.uint32(32).bool(message.saveChanges);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RefactorRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRefactorRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.oldStepValue = ProtoStepValue.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.newStepValue = ProtoStepValue.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.paramPositions.push(ParameterPosition.decode(reader, reader.uint32()));
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.saveChanges = reader.bool();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RefactorRequest {
    return {
      oldStepValue: isSet(object.oldStepValue) ? ProtoStepValue.fromJSON(object.oldStepValue) : undefined,
      newStepValue: isSet(object.newStepValue) ? ProtoStepValue.fromJSON(object.newStepValue) : undefined,
      paramPositions: globalThis.Array.isArray(object?.paramPositions)
        ? object.paramPositions.map((e: any) => ParameterPosition.fromJSON(e))
        : [],
      saveChanges: isSet(object.saveChanges) ? globalThis.Boolean(object.saveChanges) : false,
    };
  },

  toJSON(message: RefactorRequest): unknown {
    const obj: any = {};
    if (message.oldStepValue !== undefined) {
      obj.oldStepValue = ProtoStepValue.toJSON(message.oldStepValue);
    }
    if (message.newStepValue !== undefined) {
      obj.newStepValue = ProtoStepValue.toJSON(message.newStepValue);
    }
    if (message.paramPositions?.length) {
      obj.paramPositions = message.paramPositions.map((e) => ParameterPosition.toJSON(e));
    }
    if (message.saveChanges !== false) {
      obj.saveChanges = message.saveChanges;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RefactorRequest>, I>>(base?: I): RefactorRequest {
    return RefactorRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RefactorRequest>, I>>(object: I): RefactorRequest {
    const message = createBaseRefactorRequest();
    message.oldStepValue = (object.oldStepValue !== undefined && object.oldStepValue !== null)
      ? ProtoStepValue.fromPartial(object.oldStepValue)
      : undefined;
    message.newStepValue = (object.newStepValue !== undefined && object.newStepValue !== null)
      ? ProtoStepValue.fromPartial(object.newStepValue)
      : undefined;
    message.paramPositions = object.paramPositions?.map((e) => ParameterPosition.fromPartial(e)) || [];
    message.saveChanges = object.saveChanges ?? false;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}